Add font family to css
Adding multiple values separated by commas to the font-family property
in css can allow the html file to use the assigned fonts to the element
where the css is applied. Using a specific font family like "Times New
Roman" keeps the default font to the said font and using a generic
font-family name after the specific font will let the text be of the
fonts of the generic font-family if the previously assigned font is not
available to the language used.